OTP implementation on critical

+++ src/org/thothlab/devilsvault/dao/customer/ExtUserDaoImpl.java
@@ -1,10 +1,13 @@
 package org.thothlab.devilsvault.dao.customer;
+import java.math.BigInteger;
 import java.sql.Connection;
 import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
+
 import javax.sql.DataSource;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
@@ -76,7 +79,7 @@ public class ExtUserDaoImpl{
 		return balance;
      }
 	
-	public Customer setExternalUser(String name,String address,Integer phone,String email, Date date_of_birth,String ssn)
+	public Customer setExternalUser(String name,String address,BigInteger phone,String email, Date date_of_birth,String ssn)
     {
         Customer userDetails = new Customer();
         userDetails.setName(name);
@@ -103,7 +106,7 @@ public class ExtUserDaoImpl{
             ps.setString(5, "sdsdsad");
             ps.setString(6, "sdsdsad");
             ps.setInt(7, 123);
-            ps.setInt(8, userdetails.getPhone());
+            ps.setLong(8, userdetails.getPhone().longValue());
             ps.setString(9, userdetails.getEmail());
             ps.setDate(10, userdetails.getDate_of_birth());
             ps.setString(11, userdetails.getSsn());
